Better destructible trees
I repaired the trees. They were placed depending on the landscape but i modified the placement script which now allows the y axis to not be changed so they plant vertically. I also fixed collision boxes of the trees and added in a notification to say that wood has been gathered(even though i have not implemented the code to added wood to the inventory). I also fixed the clouds as they were blooming late and had poor movement.